# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: A2A Agent Platform API
  description: |
    # A2A (Agent-to-Agent) Platform REST API
    
    Enterprise-grade API for managing AI agents, workflows, and data processing pipelines.
    Built on SAP Business Technology Platform with OAuth2 security and multi-tenant support.
    
    ## Key Features
    - **Agent Management**: Create, configure, and orchestrate AI agents
    - **Workflow Automation**: Design and execute complex data processing workflows  
    - **Service Marketplace**: Discover and consume agent services
    - **Network Analytics**: Monitor performance and track metrics
    - **Security & Audit**: Enterprise-grade security with comprehensive audit trails
    
    ## Authentication
    All API endpoints require OAuth2 Bearer token authentication via SAP XSUAA.
    Scopes are enforced based on user roles and permissions defined in xs-security.json.
    
    ## Rate Limiting
    - Standard users: 1000 requests/hour
    - Premium users: 10000 requests/hour
    - Enterprise users: Unlimited
    
    ## Support
    - Documentation: https://docs.a2a-platform.com
    - Support: support@a2a-platform.com
    - Status Page: https://status.a2a-platform.com
    
  version: 1.0.0
  contact:
    name: A2A Platform Support
    email: support@a2a-platform.com
    url: https://docs.a2a-platform.com
  license:
    name: Enterprise License
    url: https://a2a-platform.com/license
  termsOfService: https://a2a-platform.com/terms
servers:
...

...
security:
  - OAuth2: []
components:
  securitySchemes:
    OAuth2:
      type: oauth2
      description: SAP XSUAA OAuth2 authentication with role-based access control
      flows:
        authorizationCode:
          authorizationUrl: https://your-tenant.authentication.sap.hana.ondemand.com/oauth/authorize
          tokenUrl: https://your-tenant.authentication.sap.hana.ondemand.com/oauth/token
          scopes:
            a2a-agents.Admin: Full administrative access to all platform functions
            a2a-agents.AgentOperator: Operate agents and manage workflows
            a2a-agents.DataViewer: View data products and reports
            a2a-agents.Agent0.Execute: Execute Agent 0 (Data Product Registration)
            a2a-agents.Agent1.Execute: Execute Agent 1 (Data Standardization)
            a2a-agents.Agent2.Execute: Execute Agent 2 (AI Data Preparation)
            a2a-agents.Agent3.Execute: Execute Agent 3 (Vector Processing)
            a2a-agents.Agent4.Execute: Execute Agent 4 (Calculation & Validation)
            a2a-agents.Agent5.Execute: Execute Agent 5 (Quality Assurance)

  # Reusable Schemas
  schemas:
    Agent:
      type: object
      description: AI Agent entity with capabilities and configuration
      required:
        - id
        - name
        - type
        - status
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the agent
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
          description: Human-readable agent name
          example: "Data Standardization Agent"
          maxLength: 100
        type:
          type: string
          enum: [data_product, standardization, ai_preparation, vector_processing, calculation, quality_assurance]
          description: Agent type determining its capabilities and role
        status:
          type: string
          enum: [active, inactive, maintenance, error]
          description: Current operational status of the agent
        version:
          type: string
          description: Agent version following semantic versioning
          example: "1.2.3"
        description:
          type: string
          description: Detailed description of agent functionality
          maxLength: 500
        capabilities:
          type: array
          description: List of agent capabilities and supported operations
          items:
            $ref: '#/components/schemas/Capability'
        configuration:
          type: object
          description: Agent-specific configuration parameters
          additionalProperties: true
        performance:
          $ref: '#/components/schemas/PerformanceMetrics'
        createdAt:
          type: string
          format: date-time
          description: Agent creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last modification timestamp
        createdBy:
          type: string
          description: User ID who created the agent
    Capability:
      type: object
      description: Agent capability definition
      required:
        - id
        - name
        - type
      properties:
        id:
          type: string
          description: Unique capability identifier
        name:
          type: string
          description: Human-readable capability name
        type:
          type: string
          enum: [data_processing, ai_inference, validation, transformation]
        parameters:
          type: object
          description: Capability-specific parameters
          additionalProperties: true
        enabled:
          type: boolean
          description: Whether capability is currently enabled
          default: true
    PerformanceMetrics:
      type: object
      description: Agent performance and health metrics
      properties:
        executionCount:
          type: integer
          description: Total number of executions
          minimum: 0
        successRate:
          type: number
          format: float
          description: Success rate percentage (0-100)
          minimum: 0
          maximum: 100
        averageExecutionTime:
          type: number
          format: float
          description: Average execution time in milliseconds
          minimum: 0
        lastExecutionTime:
          type: string
          format: date-time
          description: Timestamp of last execution
        errorCount:
          type: integer
          description: Total number of errors
          minimum: 0
        healthScore:
          type: number
          format: float
          description: Overall health score (0-100)
          minimum: 0
          maximum: 100
...
    Error:
      type: object
      description: Standard error response
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
        requestId:
          type: string
          description: Unique request identifier for tracking

  # Reusable Parameters
  parameters:
...
    Limit:
      name: limit
      in: query
      description: Maximum number of items to return
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      description: Number of items to skip
      schema:
        type: integer
        minimum: 0
        default: 0
...
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "UNAUTHORIZED"
            message: "Valid authentication token required"
            timestamp: "2024-01-15T10:30:00Z"
            requestId: "req_123456789"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "FORBIDDEN"
            message: "Insufficient permissions to access this resource"
            timestamp: "2024-01-15T10:30:00Z"
            requestId: "req_123456789"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "VALIDATION_ERROR"
            message: "Request validation failed"
            details:
              field: "name"
              issue: "Field is required"
            timestamp: "2024-01-15T10:30:00Z"
            requestId: "req_123456789"

# API Paths
paths:
...
  /agents:
    get:
      tags:
        - Agent Management
      summary: List all agents
      description: |
        Retrieve a paginated list of all agents in the platform.
        Supports filtering by status, type, and other criteria.
      operationId: listAgents
      security:
        - OAuth2: [a2a-agents.DataViewer]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: status
          in: query
          description: Filter agents by status
          schema:
            type: string
            enum: [active, inactive, maintenance, error]
        - name: type
          in: query
          description: Filter agents by type
          schema:
            type: string
            enum: [data_product, standardization, ai_preparation, vector_processing, calculation, quality_assurance]
      responses:
        '200':
          description: List of agents retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  agents:
                    type: array
                    items:
                      $ref: '#/components/schemas/Agent'
                  total:
                    type: integer
                    description: Total number of agents
                  limit:
                    type: integer
                    description: Items per page
                  offset:
                    type: integer
                    description: Items skipped
              example:
                agents:
                  - id: "550e8400-e29b-41d4-a716-446655440000"
                    name: "Data Standardization Agent"
                    type: "standardization"
                    status: "active"
                    version: "1.2.3"
                total: 25
                limit: 20
                offset: 0
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags:
        - Agent Management
      summary: Create a new agent
      description: |
        Create a new AI agent with specified capabilities and configuration.
        Requires AgentOperator or Admin permissions.
      operationId: createAgent
      security:
        - OAuth2: [a2a-agents.AgentOperator]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - type
              properties:
                name:
                  type: string
                  description: Agent name
                  maxLength: 100
                type:
                  type: string
                  enum: [data_product, standardization, ai_preparation, vector_processing, calculation, quality_assurance]
                description:
                  type: string
                  maxLength: 500
                capabilities:
                  type: array
                  items:
                    $ref: '#/components/schemas/Capability'
                configuration:
                  type: object
                  additionalProperties: true
            example:
              name: "Custom Data Processing Agent"
              type: "data_product"
              description: "Specialized agent for custom data processing workflows"
              capabilities:
                - id: "data_validation"
                  name: "Data Validation"
                  type: "validation"
                  enabled: true
      responses:
        '201':
          description: Agent created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...